[ODBC-146] The fix and the testscase.
The proble boils down to incorrect coonversion into SQL_NUMERIC in case of SQLGetData, and with binding it worked well. The patch synced the way data is converted into SQL_NUMERIC.

C/C has been updated to v_2.3.6 tag

+++ ma_parse.c
@@ -96,7 +96,7 @@ void MADB_DeleteQuery(MADB_QUERY *Query)
 int MADB_ParseQuery(MADB_QUERY * Query)
 {
   /* make sure we don't have trailing whitespace or semicolon */
-  Query->RefinedLength= SqlRtrim(Query->RefinedText, Query->RefinedLength);
+  Query->RefinedLength= SqlRtrim(Query->RefinedText, (int)Query->RefinedLength);
   Query->RefinedText=  ltrim(Query->RefinedText);
   Query->RefinedText=  FixIsoFormat(Query->RefinedText, &Query->RefinedLength);
   Query->RefinedLength-= Query->RefinedText - Query->allocated;
